refactored run_store, moved model and schema to new module
